[Zero] feat : AI 피드백 제공 로직 구현

+++ src/main/java/com/umc_9th/sleepinghero/domain/sleep/ai/SleepFeedBackContext.java
@@ -3,13 +3,13 @@ package com.umc_9th.sleepinghero.domain.sleep.ai;
 import com.umc_9th.sleepinghero.domain.sleep.dto.req.SleepReviewRequest;
 
 public record SleepFeedBackContext(
-        int sleepDuration,
-        int goalDuration,
+        long sleepDuration,
+        long goalDuration,
         int sleepStreak,
         SleepReviewRequest review
 ){
     public static SleepFeedBackContext of(
-            int sleepDuration, int goalDuration
+            long sleepDuration, long goalDuration
             , int sleepStreak, SleepReviewRequest review) {
 
         return new SleepFeedBackContext(
